+++ rts/Sim/Units/CommandAI/LineDrawer.cpp
@@ -14,7 +14,7 @@ CLineDrawer::CLineDrawer()
 void CLineDrawer::UpdateLineStipple()
 {
 	stippleTimer += (gu->lastFrameTime * cmdColors.StippleSpeed());
-	stippleTimer = fmodf(stippleTimer, (16.0f / 20.0f));
+	stippleTimer = fmod(stippleTimer, (16.0f / 20.0f));
 }
